Refactor for new design
