+++ karma.conf.js
@@ -5,7 +5,7 @@ module.exports = function(config) {
       'test/integration.test.js',
       'test/signature_validator.test.js',
       'test/keystore.test.js',
-      'test/client_service.test.js',
+      'test/client_service_with_indexeddb.test.js',
     ],
     reporters: ['progress'],
     preprocessors: {
